fix(scripts): Use a zero-width space between brackets for standup posts
Slack's markup implementation has a visual choke when parsing links from
Markdown-style formatting. When the rendering engine encounters an
opening square brace which is part of a URL definition, it treats any
other opening square braces up to the next closing square brace as part
of the link text. The following illustrates the rendering bug:

    [[This is my link text](https://google.com)]
    ^^-------------------^^ ^----------------^
    |          |          |         |
    |          |          |         `-> link target
    |          |          `-----------> link text block-end
    |          `----------------------> link text; includes open brace
    `---------------------------------> link text block-start

The opening square brace before `This` is erroneously highlighted blue
and treated as part of the link text. Adding any character between the
two block-start tokens appears to consistently produce the correct
behavior. A zero-width space is chosen purely for visual consistency.